feat: update contratos quantidade column to decimal with two-place precision

+++ src/database/migrations/20260515160437-alter-contratos-quantidade-two-decimals.js
@@ -4,7 +4,7 @@
 module.exports = {
   async up(queryInterface, Sequelize) {
     await queryInterface.changeColumn("contratos", "quantidade", {
-      type: Sequelize.DECIMAL(10, 1),
+      type: Sequelize.DECIMAL(10, 2),
       allowNull: true,
       validate: {
         min: 0,
@@ -14,7 +14,7 @@ module.exports = {
 
   async down(queryInterface, Sequelize) {
     await queryInterface.changeColumn("contratos", "quantidade", {
-      type: Sequelize.INTEGER,
+      type: Sequelize.DECIMAL(10, 1),
       allowNull: true,
       validate: {
         min: 0,
